embodiments/raspberry_pi/requirements.txt
feagi_connector
RPi.GPIO>=0.7.0,<0.8.0
opencv-python>=4.0.0,<5.0.0
